refactor: map folder structure
