+++ src/layout/HomeLayout/Home.scss
@@ -136,10 +136,6 @@
     justify-content: center;
   }
 
-  &-map {
-    padding: 0;
-  }
-
   &__content {
     align-items: center;
     padding: 0 100px;
